TestGroupAccession, supervisor testtime increased to wait more for followers to write back test results.
TestNewSupervisorElection: for robustnes testing to measure recovery ability by electing a new supervisor in case of unavaiability of the current supervisor

TestLoadTesting Experiment: Having The number of followers of a given group fixed, we invrease the messsages/sec rate and message size independently sent form followers to supervisor and measured the RTT of ACK message back to followers from supervisor.--- still implementing

ControllFollowerRole: postUIEvent(0, "[CtrlFolRole]"); line removed not to show multiple lines of , "[CtrlFolRole]" in EditBox.

SensorSpervisor Role: node.connect("server_0"); line commented out

MainActivity:  sensorSupervisorRole and sensorFollowerRole added for each node to make the node able to be the supervisor and follower of the group. This currently is used to implement newSupervisorElection Test.